requirements.txt
absl-py==0.7.1
alembic==1.1.0
altgraph==0.16.1
aniso8601==8.0.0
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
baidu-aip==2.2.17.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.7
django-cors-headers==3.1.0
django-crispy-forms==1.9.0
django-formtools==2.2
django-import-export==2.0.2
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.7
django-simple-captcha==0.5.6
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
fake-useragent==0.1.11
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.15.2
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.39
Jinja2==2.10.1
jsonschema==3.0.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lxml==4.3.0
macholib==1.11
Mako==1.1.0
Markdown==3.1
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==2.0.0
more-itertools==6.0.0
MySQL-python==1.2.5
mysqlclient==1.4.6
numpy==1.16.2
odfpy==1.4.1
openpyxl==3.0.3
pandas==0.24.2
parsel==1.5.2
pbr==5.1.3
pdfminer3k==1.3.1
pefile==2018.8.8
Pillow==5.4.1
pluggy==0.9.0
ply==3.11
protobuf==3.7.1
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyInstaller==3.4
pymongo==3.10.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyrsistent==0.15.4
pytesseract==0.2.6
pytest==4.3.1
python-dateutil==2.8.0
python-docx==0.8.10
python-editor==1.0.4
pytz==2018.9
pywin32==224
pywin32-ctypes==0.2.0
PyYAML==5.3
queuelib==1.5.0
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
Scrapy==1.7.3
seaborn==0.9.0
selenium==3.141.0
service-identity==18.1.0
six==1.10.0
soupsieve==1.8
SQLAlchemy==1.3.8
sqlparse==0.3.0
tablib==1.1.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Twisted==19.7.0
urllib3==1.24.1
w3lib==1.21.0
Werkzeug==0.15.1
wordcloud==1.5.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
